+++ src/library/model_runner-quantization.adb
@@ -472,7 +472,8 @@ package body Model_Runner.Quantization is
                 | G.Type_Q8_0 | G.Type_Q4_1 | G.Type_Q5_0 | G.Type_Q5_1
                 | G.Type_Q2_K | G.Type_Q3_K
                 | G.Type_Q4_K | G.Type_Q5_K | G.Type_Q6_K
-                | G.Type_IQ4_NL | G.Type_IQ4_XS | G.Type_MXFP4);
+                | G.Type_IQ4_NL | G.Type_IQ4_XS | G.Type_IQ3_S
+                | G.Type_MXFP4);
 
    -------------------
    -- Decode_Block --